requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
pyyaml==6.0.1; python_version >= '3.6'
